Clarify what init and last do (#59)
This was the only part of the book where I was a bit lost for a moment. While it was obvious what `last` does, I didn't know what to expect from `init`.

I think a short introduction is easier for the reader, then having to learn it from what
("Hello" :: 123 :: true :: HNil).init
returns. Especially because the rest of this section is easier to read, with the correct mental model of the target solution in mind.